pyproject.toml
[project]
name = "propertydj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.40",
    "django>=5.2.6",
    "django-ckeditor-5>=0.2.18",
    "django-environ>=0.12.0",
    "django-ninja>=1.4.3",
    "django-storages>=1.14.6",
    "pillow>=11.3.0",
    "psycopg>=3.2.10",
]
